pyproject.toml
[tool.poetry]
name = "dailydo-todo-app"
version = "0.1.0"
description = ""
authors = ["Em-Aar <mmrhaq@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
uvicorn = "^0.29.0"
sqlmodel = "^0.0.16"
psycopg = {extras = ["binary"], version = "^3.1.18"}
pytest = "^8.1.1"
httpx = "^0.27.0"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-multipart = "^0.0.9"
fastapi = "^0.111.0"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}



[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
